Make motor beeps Uno-compatible
The Arduino Uno's PWM frequency isn't as flexible as the Teensy LC so we
can't adjust the frequency to beep the motor as easily, and the library
function to do so is not in the standard Arduino API.  Instead use the
Arduino tone() function which does something similar.  It can only beep
at 50% duty cycle, but that doesn't matter much.