Add Protogasm enclosure models
These are derived from an object found on Thingiverse.  The base is
exactly the same as the original.  The lid is hacked up to remove the
header holes, add a hole for the rotary encoder, and increase the height
so the perfboard will fit.  It's not perfect but good enough for now.